refactor: Improve search functionality for sura names
Refactor the DisplayLinks component to improve the search functionality for sura names. The code now normalizes the search term and sura names to handle diacritics and Arabic characters properly. It also introduces a priority system for relevance scoring, giving higher priority to exact matches and surahs starting with the search term. This improves the accuracy of search results.

Signed-off-by: Sameh El Zankalony <zankalony@gmail.com>