feat: Add Biography and TafseerIntroPage components
The code changes add the Biography and TafseerIntroPage components to the project. These components provide the content and layout for the Biography and Tafseer Introduction pages respectively. The routes in the App component are also updated to include these new pages. This allows users to navigate to the Biography and Tafseer Introduction pages and view their content.

- Also, implemented useReducer across the website to centralize calculations as much as it could

Signed-off-by: Sameh El Zankalony <zankalony@gmail.com>