refactor: Update layout and padding in Navbar and Layout components
The code changes refactor the Navbar and Layout components to update the layout and padding. In the Navbar component, unnecessary code is removed and the flex direction is updated. In the Layout component, the padding-top value is increased to improve the spacing. These changes enhance the visual appearance and user experience of the website.

Signed-off-by: Sameh El Zankalony <zankalony@gmail.com>